requirements.txt
Django==5.1.6
psycopg2-binary>=2.9
python-decouple>=3.8
openpyxl>=3.1
Pillow>=10.0
python-dateutil>=2.9
num2words>=0.5
xhtml2pdf>=0.2.17
whitenoise>=6.5
